+++ NEWS.md
@@ -1,4 +1,19 @@
 # ospsuite.bmlm NEWS
 
+
+## [0.1.3] - 2024-02-10
+- Allow flat distribution for individual parameters
+
+## [0.1.3] - 2024-02-10
+- Bug fixes in update parameters, set default Values of missing individuals 
+
+## [0.1.2] - 2024-02-06
+- Bug fixes in update parameters, only the parameter of the first scenario are updated
+
+## [0.1.1] - 2024-02-04
+- Bug fixes in chekcProgress functions and export functions 
+  Add publix functions for class BMLMOptimization exportIndividualResultsToPkml, getCurrentConfigTable
+
+
 ## [0.1.0] - 2024-01-14
 - Initial setup of beta version